Added day 3 and debuggin support
Added a solution for day3. This is not the cleanest and most efficient
one (mod operator could have been used instead of a loop for instance),
but it can print a debugging trace (only useful with example input)

The program has a new flag, -d, which enables debugging information. Each
solution can write debugging info to a string, and, if the -d flag is
spceficied, that information will be printed